create registryable class
